gpu-market-watch/requirements.txt
requests
pydantic>=2
pyyaml
tenacity
beautifulsoup4
lxml
pandas
python-dateutil
tqdm
jinja2
